refactored modal utils
