+++ frontend/src/libs/compare-table-rows.ts
@@ -1,0 +1,21 @@
+import { Assignment } from "../api/defs/types";
+
+function compareString(str1: string, str2: string) {
+    if (str1 > str2) {
+        return 1;
+    } else if (str1 < str2) {
+        return -1;
+    }
+    return 0;
+}
+
+export function compareAssignment(a1: Assignment, a2: Assignment) {
+    return (
+        compareString(a1.position.position_code, a2.position.position_code) ||
+        compareString(
+            a1.applicant.last_name || "",
+            a2.applicant.last_name || ""
+        ) ||
+        compareString(a1.applicant.first_name, a2.applicant.first_name)
+    );
+}
