+++ frontend/src/views/admin/assignments/manipulate-assignment-confirmation.tsx
@@ -2,7 +2,43 @@ import React from "react";
 import { AdvancedFilterTable } from "../../../components/filter-table/advanced-filter-table";
 import { Button, Modal } from "react-bootstrap";
 import { Assignment } from "../../../api/defs/types";
-import { compareAssignment, assignmentModalColumn } from "./utils";
+import { compareAssignment } from "../../../libs/compare-table-rows";
+
+const assignmentModalColumn = [
+    {
+        Header: "Last Name",
+        accessor: "applicant.last_name",
+        maxWidth: 120,
+    },
+    {
+        Header: "First Name",
+        accessor: "applicant.first_name",
+        maxWidth: 120,
+    },
+    {
+        Header: "Position",
+        accessor: "position.position_code",
+        width: 200,
+    },
+    {
+        Header: "Hours",
+        accessor: "hours",
+        className: "number-cell",
+        maxWidth: 70,
+    },
+    {
+        Header: "Status",
+        maxWidth: 100,
+        id: "status",
+        // We want items with no active offer to appear at the end of the list
+        // when sorted, so we set their accessor to null (the accessor is used by react table
+        // when sorting items).
+        accessor: (data: { active_offer_status: string }) =>
+            data.active_offer_status === "No Contract"
+                ? null
+                : data.active_offer_status,
+    },
+];
 
 export function MultiManipulateOfferConfirmation(props: {
     data: Assignment[];
